rename logentry task module
