+++ src/core/tasks/__init__.py
@@ -1,5 +1,5 @@
-from core.tasks.logentry import write_admin_log
 from core.tasks.tg import send_telegram_message
+from core.tasks.write_admin_log import write_admin_log
 
 __all__ = [
     "send_telegram_message",
